backend/pyproject.toml
[project]
name = "obet"
version = "0.1.0"
description = "OBet — Prediction Market Platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.5",
    "uvicorn[standard]==0.32.0",
    "sqlalchemy==2.0.44",
    "alembic==1.17.1",
    "psycopg2-binary==2.9.11",
    "pydantic==2.10.6",
    "pydantic-settings==2.7.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "httpx==0.27.2",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",
    "email-validator==2.3.0",
]

[tool.uv]
package = false
